Added LampMultiplexer. It allows to control multiple lamps based on arbitrary conditions and itself acts as a "LampLike" object, which is basically an interface to the HueLamp.
The MotionMultiplexingLightController is a new subclass of MotionLightController, which uses the LampMultiplexer. It allows passing conditions in yaml and can control lights based on that.

What this all means: When I go to the bedroom in the night, only the night lamp turns on instead of the head lamp.
This will also be used in the living room to control both lights at once